// redux/chapter/types.ts
export enum CreateChapter {
  CREATE_CHAPTER_REQUEST = 'CREATE_CHAPTER_REQUEST',
  CREATE_CHAPTER_SUCCEEDED = 'CREATE_CHAPTER_SUCCEEDED',
  CREATE_CHAPTER_FAILED = 'CREATE_CHAPTER_FAILED',
}

export enum UpdateChapter {
  UPDATE_CHAPTER_REQUEST = 'UPDATE_CHAPTER_REQUEST',
  UPDATE_CHAPTER_SUCCEEDED = 'UPDATE_CHAPTER_SUCCEEDED',
  UPDATE_CHAPTER_FAILED = 'UPDATE_CHAPTER_FAILED',
}

export enum DeleteChapter {
  DELETE_CHAPTER_REQUEST = 'DELETE_CHAPTER_REQUEST',
  DELETE_CHAPTER_SUCCEEDED = 'DELETE_CHAPTER_SUCCEEDED',
  DELETE_CHAPTER_FAILED = 'DELETE_CHAPTER_FAILED',
}